Rework let and let! implementation
For each `let` variable in play, the Let agent now stores either a tuple
`{:todo, function, shared}` if the variable is yet to be evaluated, or a
`{:done, result}` tuple once it has been evaluated once.

Before each example is run, the agent's map is reset to
`{:todo, function, shared}` tuples according to the `Let` blocks
in context.

Other notable details:
 * The "annoying __ENV__ hack" is gone, because `let`'s `do` block is
   simply put inside a regular module function which can naturally
   access the module's other functions.
 * `let` no longer attempts to modify the `shared` state when
   evaluating a variable. As far as I can tell, the code that was trying
   to do that by storing `assigns` did not work. In any case, with `let`
   being lazily evaluated, modifying shared state would be a doubtful
   proposition.