Make let! work again (fix #86)
The let! macro was trying to insert a reference to the generated
function to force its memoization in a `before` block, but it turns
out it was just inserting an atom, not a local function call. Adding
a pair of parentheses makes the function call explicit.

However, this exposed a regression on the test introduced in c12a30c8,
requiring yet more rethought of the `let` implementation.

`let`s are now collected in a first pass, configuring their wrapper
functions (i.e. `def a, do:…` for `let a:, do: …`) to use the proper
implementations (randomly named functions) according to the current
example's context (because `let`s can be overridden, this can be
different for each example). No user code is executed during that pass.

Then, `before` blocks are run, potentially triggering the evaluation of
the `let` functions configured above.

But because `let` blocks can access to the `shared` dictionary, and
`before` blocks can modify it, we need to update the `let` blocks' view
of the `shared` dictionary after executing each `before` block. So a
copy of the `shared` dictionary is now stored and updated by the Let
agent.